pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "labscripts"
version = "0.1.3"
authors = [
    {name = "Ivan Muñoz-Gutierrez", email = "ivan.munoz.gutierrez@gmail.com"},
]
description = "Useful scripts for working with sequencing data"
readme = "README.md"
requires-python = ">=3.11"
keywords = [
    "NCBI", "Sequencing databases", "Sequencing data", "Biopython"
]
license = {text = "MIT License"}
classifiers = [
    "Programming Language :: Python",
    "License :: OSI Approved :: MIT License",
    "Intended Audience :: Science/Research",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
dependencies = [
    "biopython",
    "cgecore",
    "tabulate",
]

[project.urls]
"Homepage" = "https://github.com/ivanmugu/labscripts"

[project.scripts]
extract_sequences = "labscripts.extract_sequences.__main__:main"
fetch_sequences = "labscripts.fetch_sequences.__main__:main"
mlst = "labscripts.mlst.__main__:main"
run_mlst = "labscripts.run_mlst.__main__:main"
